added audio playing method on shlokPage(revised)
